Resueltos conflictos tras hacer pull
