+++ src/App.js
@@ -1,20 +1,22 @@
 import './App.css';
-//import Register from '../src/modules/plants/components/Register-step-2';
-//import Register from '../src/components/TaxonomicCatalog';
-import Register1 from '../src/modules/fungus/components/Register-step-1-fun';
-//import Register2 from '../src/modules/plants/components/Register-step-2';
-//import Login from '../src/modules/login/components/login';
+import { FormProvider } from '../src/modules/fungus/components/FormContext'; // Importa el FormProvider
+import RegisterStep1 from '../src/modules/fungus/components/Register-step-1-fun';
+import RegisterStep2 from '../src/modules/fungus/components/Register-step-2-fun';
+import RegisterStep3 from '../src/modules/fungus/components/Register-step-3-fun';
+
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
 function App() {
   return (
     <div className="App">
       <Router>
-        <Routes>
-          
-          <Route path="/register-step-1-fun" element={<Register1 />} />
-
-        </Routes>
+        <FormProvider> {/* Envuelve todo el Router con el FormProvider */}
+          <Routes>
+            <Route path="/register-step-1-fun" element={<RegisterStep1 />} />
+            <Route path="/register-step-2-fun" element={<RegisterStep2 />} />
+            <Route path="/register-step-3-fun" element={<RegisterStep3 />} />
+          </Routes>
+        </FormProvider> 
       </Router>
     </div>
   );
