--- src/modules/fungus/hooks/useFetchCatalog.js
@@ -1,27 +1,0 @@
-import { useEffect, useState } from 'react';
-import axios from 'axios';
-
-const useFetchCatalog = (endpoint) => {
-  const [data, setData] = useState(null);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await axios.get(`https://front-desk.onrender.com/${endpoint}`);
-        setData(response.data);
-      } catch (err) {
-        setError(err.message);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, [endpoint]);
-
-  return { data, loading, error };
-};
-
-export default useFetchCatalog;
