+++ src/modules/fungus/style/Register-step-1-fun.css
@@ -1,4 +1,3 @@
-/* src/components/Login.css */
 
 .step-1-container {
     text-align: left;
